parking/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-filter>=23.0
djangorestframework-simplejwt==5.3.1
ultralytics==8.3.43
opencv-python>=4.10.0
torch==2.6.0
torchvision==0.21.0
PyMySQL==1.1.0
cryptography>=3.4.8
requests==2.31.0
python-dotenv==1.0.0
Pillow>=10.2.0
numpy>=2.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
googlemaps>=4.10.0